requirements.txt
flask
python-dateutil
numpy
tensorflow
paddleocr
re
flask==3.0.2
json
